+++ src/parser/classes/DDBSubClass.js
@@ -252,6 +252,24 @@ export default class DDBSubClass extends DDBClass {
         icon: null,
       };
       this.data.system.advancement.push(form);
+    } else if ((this.data.name.startsWith("Circle of the Sea"))) {
+      const form = {
+        _id: foundry.utils.randomID(),
+        type: "ScaleValue",
+        configuration: {
+          distance: { units: "" },
+          identifier: `wrath-range`,
+          type: "number",
+          scale: {
+            3: { value: 5 },
+            6: { value: 10 },
+          },
+        },
+        value: {},
+        title: `Wrath Range`,
+        icon: null,
+      };
+      this.data.system.advancement.push(form);
     } else if ((this.data.name.startsWith("Psi Warrior") || this.data.name.startsWith("Soulknife")) && !this.is2014) {
       for (let advancement of this.data.system.advancement) {
         if (advancement.title !== "Energy Die") continue;